+++ lib/friendly_shipping/services/ship_engine/parse_void_response.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'friendly_shipping/bad_request'
-
 module FriendlyShipping
   module Services
     class ShipEngine
